requirements.txt
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
idna==2.10
pycodestyle==2.6.0
python-dotenv==0.15.0
requests==2.24.0
toml==0.10.2
tqdm==4.51.0
urllib3==1.25.11
